Refactor: Remove callback from writeFile in create function
The callback function in the `writeFile` call within the `create` function of `create.js` was removed. This simplifies the code as the callback was only logging the error or a "done" message, which isn't strictly necessary for the core functionality of writing data to a file. The change simplifies the code and removes an unnecessary callback.