This commit introduces two new functions in `services/creatPlayer.js`: `createPlayer` and `checkIfPlayerExist`.
`createPlayer` takes an ID and a `showPlayer` object as input. It creates a new player object, assigns it the provided ID (incremented by 1), copies the name from `showPlayer`, and retrieves the total time from `showPlayer` using `getAlltime()`.  The newly created player object is then returned.

`checkIfPlayerExist` takes an array of data and a name as input.  It iterates through the array, checking if any of the objects in the array have a `name` property that matches the provided name. If a match is found, it sets the `exist` flag to `true`. Finally, it returns the `exist` flag, indicating whether a player with the given name already exists in the data.