Add new player "david" and new riddle "math easy".
This commit adds a new player named "david" with a time of 9422 and ID 7 to the `playersDb.txt` file. It also adds a new riddle with ID 42 to the `riddle.txt` file. The new riddle is a simple math problem: "32 / 16" with the answer "2". The commit also modifies the times for existing players "elazar" and "bryan".