Refactor: Update riddle service to use API call
The `updateRiddle` function now calls the API to update a riddle instead of directly modifying an array of objects.
The function now takes no arguments.
It prompts the user for the ID of the riddle to update, the property to change, and the new value for the property.
It then makes a PUT request to the `/riddle/update/:id` endpoint with the changes and property in the request body.