Refactor: Update time to player using API call.
The `updateTimeToPlayer` function has been refactored to update the player's time using an API call to the backend instead of directly modifying the data in memory. This change ensures data persistence and consistency by updating the player's time in the database. The function now makes a PUT request to `http://localhost:2030/player/update/${idx}` with the new time, and also specifies the property to be updated as "time" in the request body.