This commit moves the `createEventToPlayer()` function call inside the "1" case of the `switch` statement in the `menu()` function.
Previously, `createEventToPlayer()` was called unconditionally at the beginning of the `menu()` function.  This meant a new player object was created regardless of whether the user chose to start a new game.  Now, a player object is only created when the user selects option "1" (to start a new game), improving efficiency by avoiding unnecessary object creation.